Removed welcome compoponent and instead implemented signin componenet directly
