+++ src/Routes.js
@@ -1,6 +1,5 @@
 import { Route, Routes } from "react-router";
 
-import Welcome from "./container/Welcome";
 import Signup from "./container/Signup";
 import Signin from "./container/Signin";
 import RequireAuth from "./RequireAuth";
@@ -9,7 +8,6 @@ export default function ApplicationRoutes() {
 
     return (
         <Routes>
-            <Route exact path="/welcome" element={<Welcome />} />
             <Route path="/welcome/signin" element={<Signin />} />
             <Route path="/welcome/signup" element={<Signup />} />
             <Route
