--- src/components/signin/styles.module.css
@@ -1,32 +1,0 @@
-.inputContainer {
-    padding-bottom: 0.5em;
-    padding-top: 0.5em;
-    position: relative;
-    display: flex;
-    align-items: center;
-}
-
-.inputContainer i {
-    position: absolute;
-    right: 1em;
-}
-
-.errorMessages {
-    text-align: left;
-    color: red;
-    font-size: 14px;
-    padding-left: 1em;
-}
-
-.inputField {
-    height: 40px !important;
-    border-radius: 0.5rem !important;
-    border: 1px solid #2A4F96;
-}
-
-.signinContainer,
-.signupContainer {
-    margin-top: 0.5rem;
-    width: 100%;
-    height: 45px;
-}
