+++ src/container/Welcome.js
@@ -1,20 +1,20 @@
-import React from "react";
-import { useNavigate } from "react-router";
+// import React from "react";
+// import { useNavigate } from "react-router";
 
-import { useAuth } from "../AuthContext";
-import WelcomeComponent from "../components/welcome";
+// import { useAuth } from "../AuthContext";
+// import WelcomeComponent from "../components/welcome/Wrapper";
 
-const Welcome = props => {
+// const Welcome = props => {
 
-    const auth = useAuth()
-    const navigate = useNavigate()
+//     const auth = useAuth()
+//     const navigate = useNavigate()
 
-    const onSignin = () =>
-        auth.signin("userOne", () => navigate("/"))
+//     const onSignin = () =>
+//         auth.signin("userOne", () => navigate("/"))
 
-    return (
-        <WelcomeComponent onSignin={onSignin} {...props} />
-    );
-}
+//     return (
+//         <WelcomeComponent onSignin={onSignin} {...props} />
+//     );
+// }
 
-export default Welcome
+// export default Welcome
